backend/requirements.txt
asgiref==3.8.1
boto3==1.34.96
botocore==1.34.96
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
jmespath==1.0.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.2
urllib3==2.2.1
